Rename extractError
This commit renames the func extractError() to extractHTTPError() and adds a new global func ExtractError().

The func extractError() is removed from internal/utclient and pkg/addrresolver nad they instead use the func ExtractError() from internal/httpauth.
